refactor : regardless to symfonyInsight analyze(#7)
