+++ composer.json
@@ -8,7 +8,8 @@
       "Config\\":"config/",
       "Model\\":"src/Model/",
       "Controller\\":"src/Controller/",
-      "Enum\\" : "src/Class/Enum"
+      "Enumeration\\" : "src/Class/Enumeration"
     }
-  }
+  },
+  "license": "ISC"
 }
